Cast rgb values to int before bit shifting to fix an implicit cast error

+++ src/Canvas/ColorUtils.php
@@ -208,15 +208,15 @@ class ColorUtils
         $backPA = $backA * (255 - $foreA);
         $pa = ($forePA + $backPA);
 
-        $b = (
+        $b = (int) (
             ($forePA * (($fore >> 8) & 0xff) + $backPA * (($back >> 8) & 0xff)) /
             $pa);
 
-        $g = (
+        $g = (int) (
             ($forePA * (($fore >> 16) & 0xff) + $backPA * (($back >> 16) & 0xff)) /
             $pa);
 
-        $r = (
+        $r = (int) (
             ($forePA * (($fore >> 24) & 0xff) + $backPA * (($back >> 24) & 0xff)) /
             $pa);
